refactor(api-server): extract schedule cadence/cron helpers from reports.rs
Behavior-preserving move of the self-contained schedule-timing concern
(cadence resolution, 5-field UNIX cron matching, HH:MM parsing, DST-aware
next-run computation) and its unit tests out of the 3.3k-line
routes/reports.rs churn hotspot into a new routes/reports/schedule_cadence.rs
submodule. reports.rs becomes reports/mod.rs; the pure helpers are moved
verbatim with visibility narrowed to pub(super) (pub(crate) re-export kept
for compute_next_run_for_schedule, used by services::scheduler). No logic
or public-path changes.
